Get indicator definitions from UIS API

Retrieves information about indicators available in the UNESCO Institute for Statistics (UIS) API, including their definitions, data availability, and associated metadata.

uis_get_indicators(
  version = NULL,
  disaggregations = FALSE,
  glossary_terms = FALSE
)

Arguments

version

Character. The API version to use. If NULL (default), the API's default version will be used.

disaggregations

Logical. If TRUE, includes disaggregation information for indicators. Default is FALSE.

glossary_terms

Logical. If TRUE, includes glossary terms associated with indicators. Default is FALSE.

Value

A data frame with information about indicators:

indicator_id

Character. The unique identifier for the indicator.

indicator_name

Character. The name of the indicator.

data_availability

List column. Contains nested information about data availability, including total record count and timeline min/max years.

entity_types

List column. Contains information about entity types associated with the indicator.

See also

Examples

# \donttest{
uis_get_indicators()
#> # A tibble: 4,247 × 7
#>    indicator_id indicator_name     theme last_data_update last_data_update_des…¹
#>    <chr>        <chr>              <chr> <chr>            <chr>                 
#>  1 10           Official entrance… EDUC… 2025-02-23       February 2025 Data Re…
#>  2 10403        Start month of th… EDUC… 2025-02-23       February 2025 Data Re…
#>  3 10404        End month of the … EDUC… 2025-02-23       February 2025 Data Re…
#>  4 10405        Start of the acad… EDUC… 2025-02-23       February 2025 Data Re…
#>  5 10406        End of the academ… EDUC… 2025-02-23       February 2025 Data Re…
#>  6 13           Theoretical durat… EDUC… 2025-02-23       February 2025 Data Re…
#>  7 200101       Total population … DEMO… 2025-02-23       February 2025 Data Re…
#>  8 200144       Population aged 1… DEMO… 2025-02-23       February 2025 Data Re…
#>  9 200151       Population aged 6… DEMO… 2025-02-23       February 2025 Data Re…
#> 10 200343       Population aged 1… DEMO… 2025-02-23       February 2025 Data Re…
#> # ℹ 4,237 more rows
#> # ℹ abbreviated name: ¹​last_data_update_description
#> # ℹ 2 more variables: data_availability <list>, entity_types <list>
# }