Get available UIS API versions

Retrieves information about the available versions of the UNESCO Institute for Statistics (UIS) API, including version IDs, descriptions, and theme data.

uis_get_versions(default = FALSE)

Arguments

default

Logical. Indicates whether only the current default version should be retrievend. Defaults to FALSE.

Value

A data frame with the following columns:

id

Character. The version identifier.

code

Character. The version code.

description

Character. The description of the version.

release_date

Character. The date when the version was released.

theme

List column. Each element is a nested data frame containing information about themes available in the version, with columns: theme_id, theme_last_update, theme_description

Examples

# \donttest{
uis_get_versions()
#> # A tibble: 6 × 4
#>   version           publication_date         description                theme   
#>   <chr>             <chr>                    <chr>                      <list>  
#> 1 20250225-2ae60fad 2025-02-27T15:29:19.309Z February 2025 DR           <tibble>
#> 2 20241121-61101499 2024-11-21T23:37:39.608Z SCCI Data refresh – Nov 2… <tibble>
#> 3 20241113-98786d81 2024-11-21T18:57:35.136Z SCCI data refresh          <tibble>
#> 4 20241030-9d4d089e 2024-10-30T17:28:00.868Z Drop data for CIV on MYS … <tibble>
#> 5 20240913-b8ca1963 2024-09-15T14:44:07.750Z Glossary Update            <tibble>
#> 6 20240910-b5ad4d82 2024-09-11T06:15:13.018Z September 2024 Data Relea… <tibble>
# }